Makefile: Add a check target and enforce it for CI and CD
Thought it would be a nice idea to automate this check :)
Unfortunately it was quite a bit more hacky than expected :o
If it's too bad, we can move the code out of the Makefile or simply
revert this commit. If not, it could make sense to add this check to the
default target (but the requirement on Xpdf might be undesirable).