Makefile: Update the URL for kki/dump.csv
The project is now also on GitHub: https://github.com/fsi-tue/kki

This also updates kki/dump.csv to the most recent version and makes sure
that the page counts remain divisible by 4.
TODO: We need to change the layout so that the long hyperlinks don't
overlap with the information on the right side.