default.nix: Switch from Xpdf to pdftk(-java)
Xpdf is dying out (in Nixpkgs it's marked as insecure and some distros
have already removed it) so it's best to use an actively maintained
replacement. This should hopefully also make the build a bit faster as
we don't need to build Xpdf anymore (but the Nix closure for pdftk has a
size of 481.8M so it might be better to use a lighter alternative -
maybe qpdf or exiftool?).

According to two CI test runs this makes the "Build the Anfiheft" step
approx. 1 minute faster (2m 29s -> 1m 35s) and reduces the paths that
need to be fetched (386.74 MiB download, 1480.64 MiB unpacked ->
255.43 MiB download, 883.60 MiB unpacked).