Updated and Shortened for WS24
We significantly shortened the content.
The How To was updated.

KKI was removed.

Adresses were updated.

Dashes for ranges, correct typesetting for time of day were updated to use en
https://de.wikipedia.org/wiki/Halbgeviertstrich#Bis-Strich

Some maps for the Sand were removed as well as the time table at the end.

Some cartoons were removed.

List of other Fachschaften and detailed info about Schwerpunkt was removed as it didn't seem relevant.