Update for WiSe 24/25 (#76)
* Updated and Shortened for WS24

We significantly shortened the content.
The How To was updated.

KKI was removed.

Adresses were updated.

Dashes for ranges, correct typesetting for time of day were updated to use en
https://de.wikipedia.org/wiki/Halbgeviertstrich#Bis-Strich

Some maps for the Sand were removed as well as the time table at the end.

Some cartoons were removed.

List of other Fachschaften and detailed info about Schwerpunkt was removed as it didn't seem relevant.

* changes SS25

* add studentische Studienberatung

* change link to short link in zdv

---------

Co-authored-by: Charlotte Meiser <63923474+Chlodde@users.noreply.github.com>
Co-authored-by: Chlodde <chlolodde@gmail.com>